+++ packages/components/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.23.10](https://github.com/GetJobber/atlantis/compare/@jobber/components@4.23.9...@jobber/components@4.23.10) (2023-07-07)
+
+**Note:** Version bump only for package @jobber/components
+
+
+
+
+
+## [4.23.9](https://github.com/GetJobber/atlantis/compare/@jobber/components@4.23.8...@jobber/components@4.23.9) (2023-07-06)
+
+**Note:** Version bump only for package @jobber/components
+
+
+
+
+
 ## [4.23.8](https://github.com/GetJobber/atlantis/compare/@jobber/components@4.23.7...@jobber/components@4.23.8) (2023-07-05)
 
 **Note:** Version bump only for package @jobber/components